refactor: tx nav item into dedicated component
shows count of txs in queue in badge
