+++ apps/atlas/src/components/navigation/header.tsx
@@ -10,24 +10,16 @@ import Sword from '@bibliotheca-dao/ui-lib/icons/sword.svg';
 import VolumeOff from '@bibliotheca-dao/ui-lib/icons/volume-mute-solid.svg';
 import VolumeOn from '@bibliotheca-dao/ui-lib/icons/volume-up-solid.svg';
 import { formatEther } from '@ethersproject/units';
-import {
-  useStarknet,
-  useStarknetTransactionManager,
-} from '@starknet-react/core';
 import Link from 'next/link';
 import { useState } from 'react';
 import { useResourcesContext } from '@/context/ResourcesContext';
 import { useAtlasContext } from '@/hooks/useAtlasContext';
 import { usePlayer } from '@/hooks/usePlayer';
-import { useWalletContext } from '@/hooks/useWalletContext';
-import { Scroll } from '@/shared/Icons';
 import NetworkConnectButton from '@/shared/NetworkConnectButton';
-import { TxStyles } from '@/shared/Validators/styles';
+import TransactionNavItem from './TransactionNavItem';
 export function Header() {
-  const { connectWallet } = useWalletContext();
   const { toggleMenuType } = useAtlasContext();
   const { lordsBalance } = useResourcesContext();
-  const { transactions } = useStarknetTransactionManager();
   const [soundOn, setSoundOn] = useState(false);
 
   const [player, currentTrack] = usePlayer([
@@ -55,12 +47,6 @@ export function Header() {
       src: '/music/scott-buckley-i-walk-with-ghosts.mp3',
     },
   ]);
-
-  const TxStyle = () => {
-    return transactions.length
-      ? TxStyles.status[transactions[0].status]
-      : 'bg-gray-200';
-  };
 
   return (
     <div className="top-0 left-0 z-40 justify-end hidden shadow-2xl bg-stone-500 sm:flex">
@@ -144,23 +130,7 @@ export function Header() {
             <Sword className="w-5 stroke-white fill-white" />
           </Button>
         </span>
-        <span>
-          <Button
-            variant="primary"
-            className="relative inline-flex"
-            onClick={() => toggleMenuType('transactionCart')}
-          >
-            <Scroll className="inline-block w-6 fill-white" />
-            <span className="flex w-3 h-3 ml-3">
-              <span
-                className={`absolute inline-flex w-3 h-3 duration-300 rounded-full opacity-75 ${TxStyle()}`}
-              ></span>
-              <span
-                className={`relative inline-flex w-3 h-3 rounded-full ${TxStyle()}`}
-              ></span>
-            </span>
-          </Button>
-        </span>
+        <TransactionNavItem />
       </div>
     </div>
   );
